fix reload list after closed model in create and delete, add delete modal

+++ src/features/user-component/user-service.ts
@@ -1,4 +1,4 @@
-import { inject, Injectable, Injector, signal } from '@angular/core';
+import { DestroyRef, inject, Injectable, Injector, signal } from '@angular/core';
 import { BaseHttpService } from '@core/services/base-http-service/base-http';
 import {
   CreateUserRequest,
@@ -8,47 +8,72 @@ import {
   UserDto,
 } from '@core/dtos';
 import { Observable } from 'rxjs';
-import { Endpoints } from '@env/endpoints';
+import { Endpoint, Endpoints } from '@env/endpoints';
 import { HttpParams } from '@angular/common/http';
 import { ModalService } from '@core/services/modal-service/modal-service';
 import { ModalDataConfiguration } from '@shared/model-ui/modal-configuration/modal-data-configuration/modal-data-configuration';
 import { UserModal } from '@features/user-component/user-modal/user-modal';
+import { DeleteModal } from '@shared/shared-ui/delete-modal/delete-modal';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Injectable()
 export class UserService {
   private readonly _baseHttpService = inject(BaseHttpService);
   private readonly _modalService = inject(ModalService);
+  private readonly _destroyRef = inject(DestroyRef);
 
   public pageParams = signal<PageRequestParams>({ page: 0, size: 10 });
   public userList = signal<UserDto[] | undefined>(undefined);
 
-  getUsers(
+  private getUsers(
     pageParams: PageRequestParams,
     extraPageParams: HttpParams,
   ): Observable<SpringPageable<UserDto>> {
     return this._baseHttpService.getPageData(Endpoints.user, pageParams, extraPageParams);
   }
 
-  getUserById(id: number): Observable<UserDto> {
+  private getUserById(id: number): Observable<UserDto> {
     return this._baseHttpService.getPageDataById(Endpoints.user, id);
   }
 
-  createUser(createUserRequest: CreateUserRequest): Observable<UserDto> {
+  public createUser(createUserRequest: CreateUserRequest): Observable<UserDto> {
     return this._baseHttpService.postData<UserDto, CreateUserRequest>(
       Endpoints.user,
       createUserRequest,
     );
   }
 
-  updateUser(updateUserRequest: UpdateUserRequest): Observable<UserDto> {
+  public updateUser(updateUserRequest: UpdateUserRequest): Observable<UserDto> {
     return this._baseHttpService.patchData<UserDto, UpdateUserRequest>(
       Endpoints.user,
       updateUserRequest,
     );
   }
 
-  deleteUser(id: number) {
-    return this._baseHttpService.deleteData(Endpoints.user, id);
+  public getAllUsers() {
+    this.getUsers(this.pageParams(), new HttpParams())
+      .pipe(takeUntilDestroyed(this._destroyRef))
+      .subscribe({
+        next: (data) => {
+          if (data) {
+            this.userList.set(data.content);
+          }
+        },
+      });
+  }
+
+  public getCurrentUserById(id: number | undefined, injector: Injector) {
+    if (id) {
+      this.getUserById(id)
+        .pipe(takeUntilDestroyed(this._destroyRef))
+        .subscribe({
+          next: (response) => {
+            if (response) {
+              this.showInfoUserModal(response, injector);
+            }
+          },
+        });
+    }
   }
 
   showCreateUserModal(injector: Injector) {
@@ -58,7 +83,17 @@ export class UserService {
       titleFallback: 'Dodaj nowego użytkownika',
     };
 
-    this._modalService.openModal(UserModal, createModalConfiguration, { injector: injector });
+    this._modalService
+      .openModal(UserModal, createModalConfiguration, {
+        injector: injector,
+      })
+      .afterClosed()
+      .pipe(takeUntilDestroyed(this._destroyRef))
+      .subscribe({
+        next: () => {
+          this.getAllUsers();
+        },
+      });
   }
 
   showInfoUserModal(response: UserDto, injector: Injector) {
@@ -72,14 +107,29 @@ export class UserService {
     this._modalService.openModal(UserModal, createModalConfiguration, { injector: injector });
   }
 
-  showDeleteModal(injector: Injector) {
-    const createModalConfiguration: ModalDataConfiguration<UserDto> = {
+  showDeleteModal(id: number | undefined, injector: Injector) {
+    if (!id) {
+      return;
+    }
+
+    const createModalConfiguration: ModalDataConfiguration<{ id: number; endpoint: Endpoint }> = {
       type: 'delete',
       title: 'MODALS.user.delete',
       titleFallback: 'Usuń użytkownika',
+      data: { id: id, endpoint: Endpoints.user },
     };
 
-    this._modalService.openModal(UserModal, createModalConfiguration, { injector: injector });
+    this._modalService
+      .openModal(DeleteModal, createModalConfiguration, {
+        injector: injector,
+      })
+      .afterClosed()
+      .pipe(takeUntilDestroyed(this._destroyRef))
+      .subscribe({
+        next: () => {
+          this.getAllUsers();
+        },
+      });
   }
 
   closeCurrentModal() {
